Get Estimated Mixture Model Coefficients

Returns the estimated coefficients of a foehnix mixture model for both, the components and the concomitant model (if specified).

# S3 method for class 'foehnix'
coef(object, type = "parameter", ...)

Arguments

object

a foehnix object.

type

character, either "parameter" "coefficients", see 'Details' section.

...

additional arguments, ignored.

Value

Returns a coef.foehnix object with the estimated coefficients.

Details

Returns the coefficients of the mixture model. If type = "parameter" the parameters of the mixture model components are returned on the 'true' scale (parameter scale), else on the link scale (as used for optimization). For each component the location and scale parameters are returned. If a concomitant model has been specified, the coefficients of the (possibly regularized) logistic regression model are returned as well (concomitant model).

See also

Author

Reto Stauffer