foehnix Demo Data Set

The foehnix package comes with two sets of meteorological observations one for Tyrol, Austria, and one for Southern California, USA. Each region comes with observations from two stations, one valley station (or target station) and one station further upstream of the main wind direction (crest station) used to filter the data (see foehnix_filter). For Tyrol, observations for station Ellbögen (valley) and station Sattelberg (crest) are included, the Californian data set consists of the crest station 'Lucky Five Ranch' and the valley station 'Viejas Casino and Resort'. More details are provided in the 'Details' section.

# Combined data set Tyrol (A).
# Valley station: Ellbögen; crest station: Sattelberg
demodata()
demodata(which = "tyrol")

# Combined data set California (CA, USA).
# Valley station: Viejas; crest station: Lucky Five Ranch
demodata(which = "california")

# Hourly observations station Ellögen (Tyrol, Austria).
# Serves as "valley" or "target" station for the "tyrol" data set.
data(which = "ellboegen")

# Hourly observations station Sattelberg (Tyrol, Austria).
# Serves as "crest" station for the "tyrol" data set.
data(which = "sattelberg")

# Hourly observations station Viejas Casino and Resort (California, U.S.A.).
# Serves as "valley" or "target" station for the "california data set.
data(which = "viejas")

# Hourly observations station Lucky Five Ranch (California, U.S.A.).
# Serves as "crest" station for the "california" data set.
data(which = "luckyfive")

demodata(which = c("tyrol", "ellboegen", "sattelberg", "california",
  "viejas", "luckyfive"))

Arguments

which
which = "tyrol" returns the combined tyrolean data set (A) and which = "california" the combined californian data set (USA), a time series object which contains measurements from both stations (valley/crest). The function can also be used to get the demo data sets for specific sites (see 'Examples'/'Usage').

Source

Station Ellögen operated by the Department of Atmospheric and Cryospheric Sciences (http://acinn.uibk.ac.at) of the University of Innsbruck. The data is available under the CC BY-SA 4.0 license.

Station Sattelberg operated by the Department of Atmospheric and Cryospheric Sciences (http://acinn.uibk.ac.at) of the University of Innsbruck. The data is available under the CC BY-SA 4.0 license.

Station Viejas Casino and Resort operated by San Diego Gas and Electric and made freely publicly available through Synoptic PBC at https://synopticdata.com/.

Station Lucky Five Ranch operated by San Diego Gas and Electric and made freely publicly available through Synoptic PBC at https://synopticdata.com/.

Details

The Tyrolean data set consists of station Elbögen located in the Wipp valley, an area well known for south foehn events. The second station called Sattelberg is located 18km (11.5mi) south of Ellbögen close to the main alpine ridge and is used as upstream crest station.

Ellbögen: 11.42889 E/47.18694 N, 1080 meters above mean sea level

Sattelberg: 11.47889 E/47.01083 N, 2107 meters above mean sea level

Hourly time series objects (class `zoo`; UTC) including the following variables:

dd: wind direction in degrees, meteorological format (0/360 correspond to 'wind from north', 90 to 'wind from east', 180 'wind from south', and 270 'wind from west')

ff: wind speed in meters per second

p: station pressure in hectopascal

rh: relative humidity in percent

t: dry air temperature in degrees Celsius

The Californian data sets consist of two stations located in the southern part of the state. Station 'Viejas' is installed next to the Viejas Casino and Resort located in Alpine, CA (valley station), the station 'Lucky Five Ranch' is located 22km (14mi) northwest of the Viejas Casino close to the main ridge of the Sierra Nevada mountain range and is used as the upstream crest station.

Viejas Casino and Resort: -116.70437 E/32.84559 N, 715 meters above mean sea level

Lucky Five Ranch: -116.528 E/32.9331 N, 1445 meters above mean sea level

Hourly time series objects (class `zoo`; UTC) including the following variables:

air_temp: dry air temperature in degrees Celsius

relative_humidity: relative humidity in percent

wind_speed: wind speed in meters per second

wind_direction: wind direction in degrees, meteorological format (0/360 correspond to 'wind from north', 90 to 'wind from east', 180 'wind from south', and 270 'wind from west')

wind_gust: wind gust speed in meters per second

See also

Examples

# Loading the combined demo data set for "Tyrol (A)".
# Stations: Ellboegen (valley station) and Sattelberg (crest station).
# Variables starting with "crest_" are observations from station Sattelberg,
x <- demodata("tyrol") # Default
print(head(x))

#>                      dd  ff  rh    t crest_dd crest_ff crest_rh crest_t diff_t
#> 2006-01-01 01:00:00 171 0.6  90 -0.4      180     10.8      100    -7.8   2.87
#> 2006-01-01 02:00:00 268 0.3 100 -1.8      186     12.5      100    -8.0   4.07
#> 2006-01-01 03:00:00 115 5.2  79  0.9      181     11.3      100    -7.4   1.97
#> 2006-01-01 04:00:00 152 2.1  88 -0.6      178     13.3      100    -7.5   3.37
#> 2006-01-01 05:00:00 319 0.7 100 -2.6      176     13.1      100    -7.1   5.77
#> 2006-01-01 06:00:00  36 0.1  99 -1.7      184     10.0      100    -6.9   5.07

# Loading the combined demo data set for "California (USA)".
# Stations: Viejas (valley station) and 'Lucky Five Ranch' (crest station).
# Variables starting with "crest_" are observations from station 'Lucky Five Ranch'.
x <- demodata("california")
print(head(x))

#>                     air_temp relative_humidity wind_speed wind_direction
#> 2012-01-01 00:00:00    23.89                25       1.34            227
#> 2012-01-01 01:00:00    20.00                22       2.68             43
#> 2012-01-01 02:00:00    19.44                15       4.02             35
#> 2012-01-01 03:00:00    20.00                13       3.58             16
#> 2012-01-01 04:00:00    20.00                12       3.58            350
#> 2012-01-01 05:00:00    18.89                13       2.68            339
#>                     wind_gust crest_air_temp crest_relative_humidity
#> 2012-01-01 00:00:00      2.68          16.67                      17
#> 2012-01-01 01:00:00      4.02          14.44                      21
#> 2012-01-01 02:00:00      6.70          13.89                      20
#> 2012-01-01 03:00:00      5.81          13.89                      19
#> 2012-01-01 04:00:00      4.47          13.89                      18
#> 2012-01-01 05:00:00      4.47          13.89                      19
#>                     crest_wind_speed crest_wind_direction crest_wind_gust
#> 2012-01-01 00:00:00             5.81                   91            8.94
#> 2012-01-01 01:00:00             0.90                   49            2.68
#> 2012-01-01 02:00:00             0.90                  154            2.68
#> 2012-01-01 03:00:00             3.13                   84            6.26
#> 2012-01-01 04:00:00             8.05                   86           10.73
#> 2012-01-01 05:00:00             8.05                   91           11.62
#>                     diff_temp
#> 2012-01-01 00:00:00      0.08
#> 2012-01-01 01:00:00      1.74
#> 2012-01-01 02:00:00      1.75
#> 2012-01-01 03:00:00      1.19
#> 2012-01-01 04:00:00      1.19
#> 2012-01-01 05:00:00      2.30

# Sattelberg only
x <- demodata("sattelberg")
print(head(x))

#>                      dd   ff  rh    t
#> 2006-01-01 01:00:00 180 10.8 100 -7.8
#> 2006-01-01 02:00:00 186 12.5 100 -8.0
#> 2006-01-01 03:00:00 181 11.3 100 -7.4
#> 2006-01-01 04:00:00 178 13.3 100 -7.5
#> 2006-01-01 05:00:00 176 13.1 100 -7.1
#> 2006-01-01 06:00:00 184 10.0 100 -6.9

# Solely Ellboegen
x <- demodata("ellboegen")
print(head(x))

#>                      dd  ff  rh    t
#> 2006-01-01 01:00:00 171 0.6  90 -0.4
#> 2006-01-01 02:00:00 268 0.3 100 -1.8
#> 2006-01-01 03:00:00 115 5.2  79  0.9
#> 2006-01-01 04:00:00 152 2.1  88 -0.6
#> 2006-01-01 05:00:00 319 0.7 100 -2.6
#> 2006-01-01 06:00:00  36 0.1  99 -1.7

# Viejas
x <- demodata("viejas")
print(head(x))

#>                     air_temp relative_humidity wind_speed wind_direction
#> 2012-01-01 00:00:00    23.89                25       1.34            227
#> 2012-01-01 01:00:00    20.00                22       2.68             43
#> 2012-01-01 02:00:00    19.44                15       4.02             35
#> 2012-01-01 03:00:00    20.00                13       3.58             16
#> 2012-01-01 04:00:00    20.00                12       3.58            350
#> 2012-01-01 05:00:00    18.89                13       2.68            339
#>                     wind_gust
#> 2012-01-01 00:00:00      2.68
#> 2012-01-01 01:00:00      4.02
#> 2012-01-01 02:00:00      6.70
#> 2012-01-01 03:00:00      5.81
#> 2012-01-01 04:00:00      4.47
#> 2012-01-01 05:00:00      4.47

# Lucky Five Ranch
x <- demodata("luckyfive")
print(head(x))

#>                     air_temp relative_humidity wind_speed wind_direction
#> 2012-01-01 00:00:00    16.67                17       5.81             91
#> 2012-01-01 01:00:00    14.44                21       0.90             49
#> 2012-01-01 02:00:00    13.89                20       0.90            154
#> 2012-01-01 03:00:00    13.89                19       3.13             84
#> 2012-01-01 04:00:00    13.89                18       8.05             86
#> 2012-01-01 05:00:00    13.89                19       8.05             91
#>                     wind_gust
#> 2012-01-01 00:00:00      8.94
#> 2012-01-01 01:00:00      2.68
#> 2012-01-01 02:00:00      2.68
#> 2012-01-01 03:00:00      6.26
#> 2012-01-01 04:00:00     10.73
#> 2012-01-01 05:00:00     11.62