Meteorological Observation Demo Data Sets

The foehnix package comes with two data sets of meteorological observations for two sites in Tyrol, Austria. One station is located nearby Ellboegen, a small town close to Innsbruck, the second station is located south of Ellboegen close to the border between Austria and Italy on top of a mountain close to the crest of the European Alps and is used as 'crest station'.

data("ellboegen")
    data("sattelberg")

Format

Returns a data.frame with a timestamp and a set of meteorological observations. See 'Details' section for more details.

Details

Sattelberg:

Location: 11.47889 E/47.01083 N

Altitude: 2107 meters above mean sea level

Ellboegen:

Location: 11.42889 E/47.18694 N

Altitude: 1080 meters above mean sea level

Variable description (columns):

timestamp: UNIX time stamp, seconds since 1970-01-01 00:00:00

dd: wind direction in degrees, meteorological format (0/360 correspond to 'wind from north', 90 to 'wind from east', 180 'wind from south', and 270 'wind from west'

ff: wind speed in meters per second

p: station pressure in hectopascal

rh: relative humidity in percent

t: dry air temperature in degrees Celsius

Source

Both stations are operated by the Department of Atmospheric and Cryospheric Sciences (http://acinn.uibk.ac.at) of the University of Innsbruck. The data is available under the CC BY-SA 4.0 license.