Fitting foehnix Mixture Model Without Concomitant Model.

Fitting method for two-component foehnix mixture model without concomitants. Typically not called directly, interfaced via foehnix.

foehnix.noconcomitant.fit(y, family, switch = FALSE, maxit = 100L,
  tol = 1e-05, verbose = TRUE, ...)

Arguments

y
numeric vector, covariate for the components of the mixture model, dimension N.

family
object of class foehnix.family.

switch
logical whether or not the two components should be switched. By default (switch = FALSE) the component which shows higher values of y is assumed to be the foehn cluster! Depending on what your covariate is you might need to switch the clusters (by setting switch = TRUE).

maxit
positive integer, or vector of length 2 with positive integer values. Maximum number of iterations of the EM algorithm. Check manual of foehnix.control for more details.

tol
numeric, or vector of length 2 containing numeric values. Tolerance for the EM algorithm. Check manual of foehnix.control for more details.

verbose
logical, default TRUE. If set to FALSE verbose output will be suppressed.

...
currently unused.

See also