Fitting Regularized foehnix Mixture Model With Concomitant Model.

Fitting method for two-component foehnix mixture model with regularization. Typically not called directly, interfaced via foehnix.

foehnix.reg.fit(
  y,
  logitX,
  family,
  glmnet.control,
  switch = FALSE,
  maxit = 100L,
  tol = 1e-05,
  verbose = TRUE,
  ...
)

Arguments

y

numeric vector, covariate for the components of the mixture model, dimension N.

logitX

numeric matrix of dimension N x P, covariates for the concomitant model (logistic regression model matrix).

family

object of class foehnix.family.

glmnet.control

an object of class glmnet.control containing the arguments for the glmnet function.

switch

logical whether or not the two components should be switched. By default (switch = FALSE) the component which shows higher values of y is assumed to be the foehn cluster! Depending on what your covariate is you might need to switch the clusters (by setting switch = TRUE).

maxit

positive integer, or vector of length 2 with positive integer values. Maximum number of iterations of the EM algorithm and the concomitant model. Check manual of foehnix.control for more details.

tol

numeric, or vector of length 2 containing numeric values. Tolerance for the optimization (EM algorithm and concomitant model). Check manual of foehnix.control for more details.

verbose

logical, default TRUE. If set to FALSE verbose output will be suppressed.

...

currently unused.

Details

TODO: Method not yet implemented, as soon as the method itself has been written: update manual page!

See also

Author

Reto Stauffer