foehnix Model Assessment Plots

Visual representation foehnix model optimization.

# S3 method for foehnix
plot(x, which = NULL, log = TRUE, ..., ask = TRUE)

Arguments

x

a foehnix mixture model object.

which

NULL (default), character, character string, integer, or numeric. Allowed characters: loglik (1), loglikcontribution (2), coef (3), hist (4), and posterior (5).

log

logical, if TRUE the x-axis for loglik, loglikcontribution and coef is shown on the log scale.

...

additional arguments, unused.

ask

boolean, default is TRUE. User will be asked to show the next figure if multiple figures are requested. Can be set to FALSE to overwrite the default.

Details

There are currently three different plot types.

"loglik" shows the log-likelihood sum path trough the iterations of the EM algorithm for parameter estimation.

"loglikcontribution" shows the log-likelihood contribution (initial value subtracted; all paths start with 0).

coef shows the development of the (standardized) coefficients during EM optimization. Parameters of the components are shown on the real scale, the coefficients of the concomitant model (if used) are shown on the standardized scale.

hist plots empirical histograms of the main variable for the two clusters (split at posterior probability >= 0.5). In addition, the estimated parametric clusters are shown.

posterior creates a histogram of the posterior probabilities. point masses at 0.0 and 1.0 indicate sharp separation of the clusters (posterior probabilities close to 0 or 1). The two additional input arguments breaks (numeric vector) and freq (logical) will be forwarded to hist(...).

Author

Reto Stauffer