foehnix Time Series Plot Config

The foehnix package comes with a default time series plotting function (see tsplot) with a pre-specified behavior regarding variable names, colors, and labels. The tsplot function, however, allows to set some custom specifications, e.g., if the names of your observations (variable name of the observations in the time series object used to train the foehnix model) are different, if different labels are required, or if you do not like our pretty colors! This function returns a control object for the time series plots for customization.

tsplot.control(...)

get.tsplot.control(x, var, property)

Arguments

...
a set of named inputs to overwrite the defaults. see 'Details' section.

var
used when calling the get.tsplot.control function. Name of the (original!) variable name

property
the property which should be returned by get.tsplot.control.

Value

Returns an object of class c("tsplot.control", "data.frame") with the specification for the time series plot.

Details

By default the tsplot function expects that the variable names are called

t: dry air temperature (degrees Celsuis)

rh: relative humidity in percent

diff_t: temperature difference to a nearby crest station

dd: wind direction in (meteorological) degrees

ff: wind speed in meters per second

ffx: gust speed in meters per second

If tsplot can find these variables, it will plot the corresponding observations, label the plots, and uses a set of default colors. The tsplot.control function allows to overrule the defaults by specifying custom values, e.g.: imagine that your wind speed variable is not ff but windspd. You can easily tell tsplot that it has to use this custom name by calling tsplot as follows:

tsplot(x, ff = "windspd")

If your wind speed variable is not even in meters per second but knots, and you dislike our default color, you can also provide a list to overwrite the default name, default color and default label by calling:

tsplot(x, ff = list(name = "windspd", color = "#0000ff", label = "wind speed in knots")

In the same way it can be used to overrule the defaults for all other variables used in the time series plotting function (see list above).