Time Series Plot of foehnix Models

Development time series plots of estimated foehnix models. TODO: they are very specific at the moment!

## Basic usage, loops trough the time series used
    ## to estimate the models.
    tsplot(x)
    ## Specify where to start/end
    tsplot(x, start, end)
    ## Change the loop-interval
    tsplot(x, ndays = ndays)
    ## Add an extra zoo object with probabilities within
    ## [0,1] for comparison
    tsplot(x, xtra = xtra)
    ## Full options
    tsplot(x, start = NULL, end = NULL, ndays = 10, ..., xtra = NULL, ask = TRUE)

Arguments

x
object of type foehnix.

start
POSIXt object or an object which can be converted to POSIXt.

end
POSIXt object or an object which can be converted to POSIXt.

ndays
integer, number of days used when looping trough the time series.

xtra
optional zoo object with probabilities in ]0,1[ to compare two classification algorithms.

ask
logical, default is TRUE.