Windrose Plot

Plotting a windrose using meteorological wind direction and wind speed (or gust speed).

Windrose plot based on a foehnix mixture model object.

Windrose plot, can handle two type of plots (see type) using the same information.

windrose(x, ...)

# S3 method for foehnix
windrose(x, type = NULL, which = NULL,
  ddvar = "dd", ffvar = "ff", mfcol = 2, maxpp = Inf, ...)

# S3 method for default
windrose(dd, ff, interval = 10, type = "density",
  windsector = NULL, main = NULL, hue = c(10, 100), power = 0.5)

Arguments

x
forwarded to windrose S3 methods.

...
forwarded as well.

type
NULL or character, one of density or histogram. If NULL both types will be plotted.

which
NULL or character, one of unconditional, nofoehn, or foehn. If NULL all three will be plotted.

ddvar
character, name of the column in the training data set which contains the wind direction information.

ffvar
character, name of the column in the training data set which contains the wind speed (or gust speed) data.

maxpp
integer (>0), maximum plots per page. Not all plots fit on one page the script loops trough.

dd
zoo object with meteorological wind directions (]0, 360[, 0/360 is 'wind from north', 90 'wind from east', 180 'wind from south', and 270 'wind from west'.

ff
zoo object with sind speed data (>= 0).

interval
numeric, a fraction of 360. If 360 cannot be devided by interval without a rest the script will stop. Interval for the segments along the wind direction dd. Default 10.

windsector
TODO: currently unused.

main
character, title. If no title is set a default one will be shown.

hue
numeric vector of length 1 or two (for type = "density" two are used, for type = "histogram" only the first one will be used at the moment). Hue(s) for colorspace::heat_hcl(...).

power
numeric >0, power parameter for the alpha channel if type = "histogram". If 1 the alpha channel is linear for the whole range of ff, values != 1 change the alpha behaviour.

x
object of class foehnix.

mfcolinteger,
number of columns of subplots.

type
character, one of density or histogram, controls the plot type.

Details

TODO: not yet finished and/or tested!

See also