Adding Wind Sector for Windrose Plots

The windrose plot allows to specify one or multiple segments to be highlighted on the plot. This is a helper function to place the segments.

windrose_add_windsector(x, ff, offset = 0.02)

Arguments

x
list containing vectors of length 2 or a matrix with two rows. If the list is named, or the matrix has row names, the names will be used to label the segments.

ff
wind speed (in case of plot type "density", or density in case of plot type "histogram". Used to draw the polygon and to plot the labels (if needed).

offset
offset for text adjustment.