Get Count Matrix for Windrose Plot

The windrose method provides one windrose/wind count plot type. This function returns a matrix with counts for different multivariate bins (binning along wind direction dd and wind speed ff).

windrose_get_counts(
  x,
  dd.breaks = seq(0, 360, by = 30),
  ff.breaks = pretty(x$ff)
)

Arguments

x

data object of type zoo or data.frame. Needs to contain at least the two columns dd (meteorological wid direction in degrees, ]0, 360[) and ff with wind speed (range >= 0).

dd.breaks

numeric vector with breaks along wind direction. Default is seq(0, 360, by = 30).

ff.breaks

numeric vector with breaks along wind speed. default is pretty(x$ff).

Value

Returns a matrix of dimension length(dd.breaks) x length(ff.breaks)

with counts >= 0.