foehnix Demo Data Set

The foehnix package comes with two demo data sets containing observations from two automated weather stations in Tyrol, Austria. One station (Sattelberg) is located close to the main alpine ridge and is used as 'crest' station. The second station (Ellboegen) is located in the Wipp valley. The package demos and examples will estimate automated foehn classification models (foehnix models) for station Ellboegen using additional information from the crest station as concomitant variables and for custom wind filters (foehnix_filter).

demodata(which = c("tyrol", "ellboegen", "sattelberg", "california",
  "viejas", "luckyfive"))

Arguments

which
the foehnix package comes with two demo data sets. Each consists of meteorological observations of two stations (one in the valley; target station) and one near the crest upstream of typical foehn direction (crest station). which = "tyrol" returns the tyrolean data set (A), which = "california" the californian data set (USA), a time series object which contains measurements from both stations (valley/crest). The function can also be used to get the demo data sets for specific sites (see 'Examples').

Source

Ellboegen and Sattelberg: Operated by the Department of Atmospheric and Cryospheric Sciences (http://acinn.uibk.ac.at) of the University of Innsbruck. The data is available under the CC BY-SA 4.0 license.

Viejas and the Lucky Five Ranch: Operated by #TODO add correct source (GM).

See also

Examples

# Loading the combined demo data set for "Tyrol (A)".
# Stations: Ellboegen (valley station) and Sattelberg (crest station).
# Variables starting with "crest_" are observations from station Sattelberg,
x <- demodata("tyrol") # Default
print(head(x))

#>                      dd  ff  rh    t crest_dd crest_ff crest_rh crest_t diff_t
#> 2006-01-01 01:00:00 171 0.6  90 -0.4      180     10.8      100    -7.8   2.87
#> 2006-01-01 02:00:00 268 0.3 100 -1.8      186     12.5      100    -8.0   4.07
#> 2006-01-01 03:00:00 115 5.2  79  0.9      181     11.3      100    -7.4   1.97
#> 2006-01-01 04:00:00 152 2.1  88 -0.6      178     13.3      100    -7.5   3.37
#> 2006-01-01 05:00:00 319 0.7 100 -2.6      176     13.1      100    -7.1   5.77
#> 2006-01-01 06:00:00  36 0.1  99 -1.7      184     10.0      100    -6.9   5.07

# Loading the combined demo data set for "California (USA)".
# Stations: Viejas (valley station) and 'Lucky Five Ranch' (crest station).
# Variables starting with "crest_" are observations from station 'Lucky Five Ranch'.
x <- demodata("california")
print(head(x))

#>                     air_temp relative_humidity wind_speed wind_direction
#> 2012-01-01 00:00:00    23.89                25       1.34            227
#> 2012-01-01 01:00:00    20.00                22       2.68             43
#> 2012-01-01 02:00:00    19.44                15       4.02             35
#> 2012-01-01 03:00:00    20.00                13       3.58             16
#> 2012-01-01 04:00:00    20.00                12       3.58            350
#> 2012-01-01 05:00:00    18.89                13       2.68            339
#>                     wind_gust crest_air_temp crest_relative_humidity
#> 2012-01-01 00:00:00      2.68          16.67                      17
#> 2012-01-01 01:00:00      4.02          14.44                      21
#> 2012-01-01 02:00:00      6.70          13.89                      20
#> 2012-01-01 03:00:00      5.81          13.89                      19
#> 2012-01-01 04:00:00      4.47          13.89                      18
#> 2012-01-01 05:00:00      4.47          13.89                      19
#>                     crest_wind_speed crest_wind_direction crest_wind_gust
#> 2012-01-01 00:00:00             5.81                   91            8.94
#> 2012-01-01 01:00:00             0.90                   49            2.68
#> 2012-01-01 02:00:00             0.90                  154            2.68
#> 2012-01-01 03:00:00             3.13                   84            6.26
#> 2012-01-01 04:00:00             8.05                   86           10.73
#> 2012-01-01 05:00:00             8.05                   91           11.62
#>                     diff_temp
#> 2012-01-01 00:00:00      0.08
#> 2012-01-01 01:00:00      1.74
#> 2012-01-01 02:00:00      1.75
#> 2012-01-01 03:00:00      1.19
#> 2012-01-01 04:00:00      1.19
#> 2012-01-01 05:00:00      2.30

# Sattelberg only
x <- demodata("sattelberg")
print(head(x))

#>                      dd   ff  rh    t
#> 2006-01-01 01:00:00 180 10.8 100 -7.8
#> 2006-01-01 02:00:00 186 12.5 100 -8.0
#> 2006-01-01 03:00:00 181 11.3 100 -7.4
#> 2006-01-01 04:00:00 178 13.3 100 -7.5
#> 2006-01-01 05:00:00 176 13.1 100 -7.1
#> 2006-01-01 06:00:00 184 10.0 100 -6.9

# Solely Ellboegen
x <- demodata("ellboegen")
print(head(x))

#>                      dd  ff  rh    t
#> 2006-01-01 01:00:00 171 0.6  90 -0.4
#> 2006-01-01 02:00:00 268 0.3 100 -1.8
#> 2006-01-01 03:00:00 115 5.2  79  0.9
#> 2006-01-01 04:00:00 152 2.1  88 -0.6
#> 2006-01-01 05:00:00 319 0.7 100 -2.6
#> 2006-01-01 06:00:00  36 0.1  99 -1.7

# Viejas
x <- demodata("viejas")
print(head(x))

#>                     air_temp relative_humidity wind_speed wind_direction
#> 2012-01-01 00:00:00    23.89                25       1.34            227
#> 2012-01-01 01:00:00    20.00                22       2.68             43
#> 2012-01-01 02:00:00    19.44                15       4.02             35
#> 2012-01-01 03:00:00    20.00                13       3.58             16
#> 2012-01-01 04:00:00    20.00                12       3.58            350
#> 2012-01-01 05:00:00    18.89                13       2.68            339
#>                     wind_gust
#> 2012-01-01 00:00:00      2.68
#> 2012-01-01 01:00:00      4.02
#> 2012-01-01 02:00:00      6.70
#> 2012-01-01 03:00:00      5.81
#> 2012-01-01 04:00:00      4.47
#> 2012-01-01 05:00:00      4.47

# Lucky Five Ranch
x <- demodata("luckyfive")
print(head(x))

#>                     air_temp relative_humidity wind_speed wind_direction
#> 2012-01-01 00:00:00    16.67                17       5.81             91
#> 2012-01-01 01:00:00    14.44                21       0.90             49
#> 2012-01-01 02:00:00    13.89                20       0.90            154
#> 2012-01-01 03:00:00    13.89                19       3.13             84
#> 2012-01-01 04:00:00    13.89                18       8.05             86
#> 2012-01-01 05:00:00    13.89                19       8.05             91
#>                     wind_gust
#> 2012-01-01 00:00:00      8.94
#> 2012-01-01 01:00:00      2.68
#> 2012-01-01 02:00:00      2.68
#> 2012-01-01 03:00:00      6.26
#> 2012-01-01 04:00:00     10.73
#> 2012-01-01 05:00:00     11.62