foehnix Demo Data Set

The foehnix package comes with two demo data sets containing observations from two automated weather stations in Tyrol, Austria. One station (Sattelberg) is located close to the main alpine ridge and is used as 'crest' station. The second station (Ellboegen) is located in the Wipp valley. The package demos and examples will estimate automated foehn classification models (foehnix models) for station Ellboegen using additional information from the crest station as concomitant variables and for custom wind filters (foehnix_filter).

demodata(which = "combined")

Arguments

which
either "combined" (returns a combined data set), "sattelberg" (only observations from station Sattelberg), or "ellboegen" (only observations rom station Ellboegen).

Source

Station operated by the Department of Atmospheric and Cryospheric Sciences (http://acinn.uibk.ac.at) of the University of Innsbruck. The data is available under the CC BY-SA 4.0 license.

See also

Examples

# Loading the combined demo data set. Variables starting with
# "crest_" are observations from station Sattelberg.
x <- demodata()
print(head(x))

#>                      dd  ff     p rh    t crest_dd crest_ff crest_p crest_rh
#> 2010-01-18 09:00:00 338 1.2 899.8 75 -0.9        3      5.6   787.7       93
#> 2010-01-18 10:00:00  18 0.1 899.9 66  0.2      344      5.3   787.9       90
#> 2010-01-18 11:00:00 344 0.8 899.1 63  0.7      360      6.0   787.3       80
#> 2010-01-18 12:00:00 303 2.4 898.8 60  0.7      357      6.5   786.8       77
#> 2010-01-18 13:00:00 292 2.8 898.5 57  1.9      358      8.7   786.6       78
#> 2010-01-18 14:00:00 316 1.6 898.6 70  0.3      352      7.6   786.8       80
#>                     crest_t diff_t
#> 2010-01-18 09:00:00    -8.7   2.47
#> 2010-01-18 10:00:00    -8.0   2.07
#> 2010-01-18 11:00:00    -7.7   1.87
#> 2010-01-18 12:00:00    -7.3   2.27
#> 2010-01-18 13:00:00    -7.2   1.17
#> 2010-01-18 14:00:00    -7.2   2.77

# Sattelberg only
x <- demodata("sattelberg")
print(head(x))

#>                      dd  ff     p rh    t
#> 2010-01-18 09:00:00   3 5.6 787.7 93 -8.7
#> 2010-01-18 10:00:00 344 5.3 787.9 90 -8.0
#> 2010-01-18 11:00:00 360 6.0 787.3 80 -7.7
#> 2010-01-18 12:00:00 357 6.5 786.8 77 -7.3
#> 2010-01-18 13:00:00 358 8.7 786.6 78 -7.2
#> 2010-01-18 14:00:00 352 7.6 786.8 80 -7.2

# Ellboegen only
x <- demodata("ellboegen")
print(head(x))

#>                      dd  ff     p rh    t
#> 2010-01-18 09:00:00 338 1.2 899.8 75 -0.9
#> 2010-01-18 10:00:00  18 0.1 899.9 66  0.2
#> 2010-01-18 11:00:00 344 0.8 899.1 63  0.7
#> 2010-01-18 12:00:00 303 2.4 898.8 60  0.7
#> 2010-01-18 13:00:00 292 2.8 898.5 57  1.9
#> 2010-01-18 14:00:00 316 1.6 898.6 70  0.3