Foehn Classification Based on a Two-Component Mixture Model

This is the main method of the foehnix package to estimate two-component mixture models for automated foehn classification.

The default image plot of a foehnix object is a Hovmoeller diagram.

foehnix(formula, data, switch = FALSE, filter = NULL,
  family = "gaussian", control = foehnix.control(family, switch, ...),
  ...)

# S3 method for foehnix
logLik(object, ...)

# S3 method for foehnix
AIC(object, ...)

# S3 method for foehnix
BIC(object, ...)

# S3 method for foehnix
edf(object, ...)

# S3 method for foehnix
print(x, ...)

# S3 method for foehnix
formula(x, ...)

# S3 method for foehnix
summary(object, ...)

# S3 method for foehnix
image(x, FUN = "freq", deltat = NULL, deltad = 7L,
  col = rev(gray.colors(20)), contours = FALSE,
  contour.col = "black", zlim = NULL, ...)

Arguments

formula
an object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under 'Details'.

data
a time series object of class zoo containing the variables for the two-part mixture model.

switch
logical. If set to TRUE the two estimated components will be switched. This is important if the covariate for the components (left hand side of input formula is smaller for cases with foehn than for cases without foehn, e.g., when using the temperature difference as main covariate.

filter
a named list can be provided to apply a custom (simple) filter to the observations on data. Can be used to e.g., prespecify a specific wind sector for foehn winds. Please read the manual of the foehnix_filter method for more details and examples.

family
character (at the moment "gaussian" or "logistic") or an object of class foehnix.family.

control
additional control arguments, see foehnix.control.

...
forwarded to foehnix.control

x
object of class foehnix.

FUN
character string or a custom aggregation function. See 'Details' section for more information.

deltat
integer, interval in seconds for the time of day axis. Has to be a fraction of 86400 (24 hours in seconds). It NuLL (default) the interval of the time series object will be used.

col
vector of colors forwarded to image.default. By default a gray scale color map is used.

contours
logical TRUE or FALSE, whether or not the concours should be added.

contour.col
color for the contour lines, only used if contours = TRUE.

zlim
vector of two numeric values to manually specify the range of the color map used.

deltat
integer, similar to deltat, the interval in days for the grid on the x-axis. Default is 7L (aggregate to weekly values).

...
additional arguments forwarded to the image and contour plot.

Value

Returns an object of class foehnix.

Details

The two-component mixture model can be specified via formula object where the left hand side of the formula contains the variable explaining the two components (only one variable), the right hand side of the formula specifies the concomitant variables (multiple variables allowed). As an example: let's assume that our zoo object 'data' contains the following columns:

ff: observed wind speed at target site

rh: observed relative humidity at target site

diff_t: temperature difference between target site and a station upstream of the foehn wind direction

The specification for formula could e.g. look as follows:

ff ~ 1: the two components of the mixture model will be based on the observed wind speed (ff), no concomitant variables are given (intercept only).

ff ~ rh: similar to the specification above, but using observed relative humidity (rh) as concomitant.

ff ~ rh + diff_t: as above but with an additional second concomitant variable (observed temperature difference, diff_t).

diff_t ~ ff + rh: using temperature difference as the main variable for the two components while ff and rh are used as concomitants.

Note that these are just examples and have to be adjusted given data availability, location, structure/names of the variables in the data object.

An optional numeric vector of length 2 can be provided to filter the data. If set, only times (rows in data) where the observed wind direction dd lies within the wind sector specified. For the examples the wind direction dd is assumed to be in degrees, meteorological definition (90 = from east, 180 = from south, …). The following condition is used:

If windsector[1L] < windsector[2L]: use data where data$dd >= windsector[1L] & data$dd <= windsector[2L] (e.g., windsector = c(90, 270) for south wind sector).

If windsector[1L] > windsector[2L]: use data where data$dd >= windsector[1L] | data$dd <= windsector[2L] (e.g., windsector = c(270, 90) for north wind sector).

TODO: Details for lambda. If "auto" the default is "loglik" except that the correlation between two or more covariates used in formula exceeds 0.75 (both positive or negative). In this case lambda.min = "AIC" will be used automatically with additional regularization. If lambda.min = "loglik" an unpenalized estimate will be returned. If lambda.min is either AIC or BIC a ridge penalty (L2 penalty) is used. A fixed set of penalties will be tested. The one minimising either the AIC or BIC criteria will be used for the foehn classification model.

Plotting a Hovmoeller diagram based on the zoo time series object of the foehnix classification. Different plot types are available. The default functions (see list below) use na.rm = TRUE.

Input FUN can be one of the following character strings:

mean: mean probability.

freq: plotting frequencies.

occ: plotting occurance of foehn (where probability >= 0.5).

noocc: contrary to occ: occurance of no foehn (probability < 0.5).

FUN can also be a custom function used for time series aggregation (see aggregate.zoo).

References

Plavcan D, Mayr GJ, Zeileis A (2014). Automatic and Probabilistic Foehn Diagnosis with a Statistical Mixture Model. Journal of Applied Meteorology and Climatology. 53(3), 652--659. doi: 10.1175/JAMC-D-13-0267.1

Gr\"un B, Leisch F (2007). Fitting Finite Mixtures of Generalized Linear Regressions in R. Computational Statistics \& Data Analysis. 51(11), 5247--5252. doi: 10.1016/j.csda.2006.08.014

Gr\"un B, Leisch F (2008). FlexMix Version 2: Finite Mixtures with Concomitant Variables and Varying and Constant Parameters. Journal of Statistical Software, Articles. 28(4), 1--35. doi: 10.18637/jss.v028.i04

Fraley C, Raftery AE (2002). Model-Based Clustering, Discriminant Analysis, and Density Estimation. Journal of the American Statistical Association. 97(458), 611--631. doi: 10.1198/016214502760047131

See also