foehnix Two-Component Mixture-Model Control Object

Used to control the foehnix mixture models.

foehnix.control(family, switch, left = -Inf, right = Inf,
  truncated = FALSE, standardize = TRUE, maxit = 100L, tol = 1e-08,
  alpha = NULL, verbose = TRUE, ...)

Arguments

family
character specifying the distribution of the components in the mixture model. Allowed: "gaussian" and "logistic". For experts: custom foehnix.family objects can be provided as well.

switch
logical whether or not the two components should be switched. By default (switch = FALSE) the component which shows higher values of y is assumed to be the foehn cluster! Depending on what your covariate is you might need to switch the clusters (by setting switch = TRUE).

left
default is -Inf, left censoring or truncation point. See also input right and input truncated. Can be set to any finite numeric value.

right
default is Inf, right censoring or truncation point. See also input left and input truncated. Can be set to any finite numeric value.

truncated
logical. If set to TRUE truncation is used instead of censored. This only affects the foehnix model estimate if input left and/or input right are specified.

standardize
logical flag, default is TRUE. Defines whether or not the model matrix for the concomitant model should be standardized for model estimation. Recommended.

maxit
control argument for the iterative solvers. Default is 100L, the maximum number of iterations for the EM algorithm and the IWLS backfitting algorithm for the concomitant model. If a vector of length two is provided the first value is used for the EM algorithm, the second for the IWLS backfitting.

tol
similar as for maxit. Used to identify convergence of the iterative solvers. Default is 1e-8, if two values are given the first will be used for the EM algorithm, the second one for the IWLS backfitting procedure. If set to -Inf maxit will be used as stopping criteria.

alpha
TODO alpha parameter for the penalization of the concomitant model.

verbose
logical, if set to FALSE output is suppressed.

...
currently set to hell.

See also