Draw Windrose

#TODO: Note that this is, at the moment, just a copy
    # of an existing method I've written some time ago and is not
    # ment to enter the release as it is.
    windrose(x, interval = 30, ff = NULL, circle = seq(0,1,by=0.02),
             windsextor = NULL, main = NULL, draw.gusts = FALSE )

Arguments

x
zoo or data.frame object containing at least the two columns dd and ff.

interval
numeric or integer, interval in degrees for the binning of the data.

ff
default NULL … maybe name of the wind speed column in x? TODO: no idea.

circle
sequence where the circles should be drawn.

windsector
NULL by default, a wind sector can be provided (as for foehnix) to highlight the sector from which foehn winds are observed.

main
NULL by default or a character containing the title for the windrose.

draw.gusts
logical. If TRUE the function expects to find ffx in the input data set x.

Details

This function requires quite some work! As it is right now it is simply a copy of an existing method I've written a while ago.